+++ restclient.go
@@ -5,6 +5,7 @@ import (
 	"io"
 	"io/ioutil"
 	"net/http"
+	"time"
 
 	"github.com/pkg/errors"
 )
@@ -20,13 +21,34 @@ type RESTRequester interface {
 // The togglRESTAPIClient perform the HTTP requests against the Toggl API and
 // returns the APIs' response.
 type togglRESTAPIClient struct {
-	baseURL string
-	token   string
+	baseURL              string
+	token                string
+	pauseBetweenRequests time.Duration // e.g. time.Millisecond * 1000
+
+	lastRequestTimestamp time.Time
 }
 
 // Request sends an HTTP request with the given parameters (method, route, payload) to the Toggl
 // API and returns the APIs' response or an error if the request failed.
 func (client *togglRESTAPIClient) Request(method, route string, payload io.Reader) ([]byte, error) {
+
+	// pause between requests to make sure not
+	// more than ~ one request per second.
+	timeSinceLastRequest := time.Since(client.lastRequestTimestamp)
+	if timeSinceLastRequest < client.pauseBetweenRequests {
+		waitTime := client.pauseBetweenRequests - timeSinceLastRequest
+		time.Sleep(waitTime)
+	}
+
+	// capture the request time
+	client.lastRequestTimestamp = time.Now()
+
+	return client.request(method, route, payload)
+}
+
+// request sends an HTTP request with the given parameters (method, route, payload) to the Toggl
+// API and returns the APIs' response or an error if the request failed.
+func (client *togglRESTAPIClient) request(method, route string, payload io.Reader) ([]byte, error) {
 
 	httpClient := &http.Client{}